requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Redis
redis[hiredis]==5.0.1

# Configuration and validation
pydantic==2.4.2
pydantic-settings==2.0.3

# Security
argon2-cffi==23.1.0
python-multipart==0.0.6

# HTTP client for provider APIs
httpx==0.25.2

# Token counting
tiktoken>=0.7.0

# Monitoring and metrics
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
boto3>=1.28.0

# OpenTelemetry Tracing
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0
setuptools<70.0.0